refactor: factorize some functions
