+++ challenge/internal/client_error_test.go
@@ -1,4 +1,4 @@
-package dns01
+package internal
 
 import (
 	"errors"
@@ -9,9 +9,9 @@ import (
 )
 
 func TestDNSError_Error(t *testing.T) {
-	msgIn := createDNSMsg("example.com.", dns.TypeTXT, true)
+	msgIn := CreateDNSMsg("example.com.", dns.TypeTXT, true)
 
-	msgOut := createDNSMsg("example.org.", dns.TypeSOA, true)
+	msgOut := CreateDNSMsg("example.org.", dns.TypeSOA, true)
 	msgOut.Rcode = dns.RcodeNameError
 
 	testCases := []struct {
